requirements.txt
Flask==3.0.0
twilio==8.11.0
python-dotenv==1.0.0
